Re-arrange repository as required and rename main code to forward
We have to put the module in a sub-folder for this to work properly
with pip.

I've renamed the main code to forward as I think that makes most sense.
FAIR is then called with 'fair.forward(input arguments)' which is clear
enough and better than 'fair_main_pub.py'. Happy to change the name if
we want though.