Fix up q values
The default q values didn't line up with the default values given
in the paper. I have also added default TCR and ECS values (as I
think this makes more sense) and added a line in the commenting
explaining that if TCR/ECS values are supplied than any q values
that are also supplied will be overwritten.