Add CH4 and N2O into the model
Have added Methane and N2O into the model in as basic a way as possible with fixed lifetimes. I've defined a few more functions in order to keep the code from being too repetitive or long, but on sections that I've majorly altered, I've just commented out the previous code so reverting back if desired is easier.