Indentation, bugs, added RF out, changed concentration definitions
Changed indentation to be consistent throughout.
Altered the RF formulae for CH4 and N2O- IPCC report has typo which I found by checking the source (Myhre et al).
Added RF as an output this is often used as a metric by IPCC so I found it useful to be able to check that the expected radiative forcings match up with the code.
Changed M and N such that they now represent total concentration (ie. anomaly + unperturbed). This was to ensure that having concentration as an input works, and carbon was already done in this way so this is more consistent.