Add CH4 and N2O into plotting function, commiting RCP file including CH4 and N2O
Added in CH4 and N2O into plot_fair. Done this identically to Carbon, which has lengthened the function quite a bit (more parameters) but not added any more code really. Done like this as working with the assumption that if this version of FAIR is used they want CH4 and N2O to be included so have not made backwards compatible (it will always display CH4 and N2O graphs).