Additions of CFCs/HCFCs/HFCs
I've added in a selection (to be expanded if desired) of other GHGs into the model.

Since all the other GHGs have radiative forcing formulae in the same format, it made sense to treat them all together by storing their variables in multidimensional arrays, that can then be treated identically to as if there were only one gas.

The only slightly tricky bit that comes out of this is the format of the output. The computation requires that the 2D array containing all the extra gas concentrations be such that time is the higher dimension to be looped over. However, for the output it is desirable that the gas species are the higher output- so they are separate. This is have done by adding on a swapaxes statement.

I also altered the format of some of the concentration computation lines to make them clearer.